Adds power command
